Possible fix for hang with Intel MPI
Was getting intermittent hangs with Intel MPI. May be due to
having no MPI calls in the main loop if all ranks other than
zero have finished their jobs. Modified to keep calling
MPI_Testsome in this case in the hope that this will help
to ensure progression (not clear why this is necessary).